pyproject.toml
[project]
name = "str-cirurgia"
version = "0.1.0"
description = "Thiago's Professional Page"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.7.0",
    "fpdf2>=2.8.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pycryptodome>=3.21.0",
    "streamlit>=1.39.0",
    "pymongo==3.11.0",
    "streamlit-authenticator==0.2.3",
]
